Don't use cf ssh -N as it confuses people
We ran a usability testing session to see if people were able to follow these
instructions to access a postgres service and import a database dump.

3 people, experienced developers and ops engineers, got as far as this step and
all of them were confused into thinking that the cf ssh command was hanging
because they didn't see a prompt or any other confirmation that the connection
succeeded.

The behaviour with -N is similar to that when you ssh interactively to a port
that is firewalled and dropping packets, so that seems like a likely
explanation for it.

There doesn't seem to be any advantage to using the -N flag in this scenario,
so I removed it entirely.